ヘッダがモーションの最初に表示されるのを解消
