+++ frontend/src/components/homepage/HomePage.css
@@ -99,7 +99,7 @@ body {
 
 .hero-logo {
   position: absolute;
-  top: 45%;
+  top: 46%;
   transform: translate(-50%, -50%);
   font-size: 3rem;
   color: white;
