+++ frontend/src/components/layout/Header.jsx
@@ -5,7 +5,7 @@ import { useAuth } from '../../hooks/AuthContext';
 const Header = () => {
   const { user } = useAuth();
   return (
-    <div className="fixed top-0 left-0 w-full bg-stone-600 shadow-md z-50 p-4 h-16">
+    <div className="fixed top-0 left-0 w-full bg-stone-600 shadow-md z-50 p-4 h-14">
       { user ? <LoginHeader /> : <NotLoginHeader />}
     </div>
   )
